+++ packages/popover/src/use-popover.ts
@@ -18,12 +18,15 @@ import {
   contains,
   FocusableElement,
   getOwnerDocument,
-  HTMLProps,
   isBrowser,
-  PropGetter,
   px,
 } from "@chakra-ui/utils"
-import { mergeRefs, getRelatedTarget } from "@chakra-ui/react-utils"
+import {
+  mergeRefs,
+  getRelatedTarget,
+  HTMLProps,
+  PropGetter,
+} from "@chakra-ui/react-utils"
 import { RefObject, useCallback, useEffect, useRef } from "react"
 
 const TRIGGER = {
